+++ src/main/java/org/folio/pv/client/PwnedClientAddPaddingInterceptor.java
@@ -1,23 +1,27 @@
 package org.folio.pv.client;
 
-import feign.RequestInterceptor;
-import feign.RequestTemplate;
+import java.io.IOException;
+import org.springframework.http.HttpRequest;
+import org.springframework.http.client.ClientHttpRequestExecution;
+import org.springframework.http.client.ClientHttpRequestInterceptor;
+import org.springframework.http.client.ClientHttpResponse;
 
-public class PwnedClientAddPaddingInterceptor implements RequestInterceptor {
+public class PwnedClientAddPaddingInterceptor implements ClientHttpRequestInterceptor {
 
   private static final String ADD_PADDING_HEADER = "Add-Padding";
 
   private final boolean paddingEnabled;
-
 
   public PwnedClientAddPaddingInterceptor(boolean paddingEnabled) {
     this.paddingEnabled = paddingEnabled;
   }
 
   @Override
-  public void apply(RequestTemplate template) {
+  public ClientHttpResponse intercept(HttpRequest request, byte[] body, ClientHttpRequestExecution execution)
+    throws IOException {
     if (paddingEnabled) {
-      template.header(ADD_PADDING_HEADER, "true");
+      request.getHeaders().add(ADD_PADDING_HEADER, "true");
     }
+    return execution.execute(request, body);
   }
 }