+++ src/main/java/org/folio/pv/service/validator/ValidatorRegistryImpl.java
@@ -1,6 +1,5 @@
 package org.folio.pv.service.validator;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
@@ -8,8 +7,9 @@ import org.folio.pv.client.PwnedClient;
 import org.folio.pv.domain.RuleType;
 import org.folio.pv.domain.entity.PasswordValidationRule;
 import org.folio.spring.FolioExecutionContext;
-import org.springframework.lang.NonNull;
+import org.jspecify.annotations.NonNull;
 import org.springframework.stereotype.Component;
+import tools.jackson.databind.ObjectMapper;
 
 @Component
 @RequiredArgsConstructor