+++ src/test/java/org/folio/pv/service/PasswordValidatorServiceImplTest.java
@@ -17,6 +17,8 @@ import io.github.glytching.junit.extension.random.RandomBeansExtension;
 import java.util.Optional;
 import java.util.UUID;
 import org.folio.pv.client.UserClient;
+import org.folio.pv.client.model.User;
+import org.folio.pv.domain.RuleType;
 import org.folio.pv.domain.ValidationType;
 import org.folio.pv.domain.dto.Password;
 import org.folio.pv.domain.dto.PasswordCheck;
@@ -27,6 +29,7 @@ import org.folio.pv.service.exception.NoRulesMatchedException;
 import org.folio.pv.service.exception.UserNotFoundException;
 import org.folio.pv.service.validator.RegExpValidator;
 import org.folio.pv.service.validator.ValidatorRegistry;
+import org.folio.spring.testing.type.UnitTest;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -35,7 +38,10 @@ import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.client.HttpClientErrorException;
 
+@UnitTest
 @ExtendWith({MockitoExtension.class, RandomBeansExtension.class})
 class PasswordValidatorServiceImplTest {
   private static final String INVALID_PASSWORD = "password.invalid";
@@ -107,6 +113,7 @@ class PasswordValidatorServiceImplTest {
       @Random PasswordValidationRule enabledRule) {
     String tenant = UUID.randomUUID().toString();
     enabledRule.setName(ruleName);
+    enabledRule.setRuleType(RuleType.REGEXP);
     mockValidatorByRule(tenant, enabledRule);
     ValidationErrors errors = ValidationErrors.none();
     mockValidator(passwordCheck, errors);
@@ -122,9 +129,10 @@ class PasswordValidatorServiceImplTest {
   )
   void checkPassword_shouldFailWithValidatorMsg(String ruleName, @Random PasswordCheck passwordCheck,
       @Random PasswordValidationRule enabledRule) {
-    String tenant = UUID.randomUUID().toString();
     enabledRule.setName(ruleName);
+    enabledRule.setRuleType(RuleType.PROGRAMMATIC);
     enabledRule.setValidationType(ValidationType.STRONG);
+    String tenant = UUID.randomUUID().toString();
     when(validationRuleService.getEnabledRules(tenant)).thenReturn(singletonList(enabledRule));
 
     NoRulesMatchedException exception = Assertions.assertThrows(NoRulesMatchedException.class,
@@ -133,8 +141,24 @@ class PasswordValidatorServiceImplTest {
     assertThat(exception.getMessage()).containsIgnoringCase("No matched rules");
   }
 
+  @Test
+  void shouldCatchHttpNotFoundExceptionWhenUserNotFound(@Random String tenant, @Random Password password) {
+    String userId = password.getUserId();
+    when(userClient.getUserById(userId))
+      .thenThrow(HttpClientErrorException.NotFound.create(
+        HttpStatus.NOT_FOUND, "User not found", null, null, null
+      ));
+
+    UserNotFoundException exc = Assertions.assertThrows(UserNotFoundException.class,
+      () -> service.validatePasswordByRules(tenant, password));
+
+    assertAll(
+      () -> assertThat(exc.getMessage()).containsIgnoringCase("not found"),
+      () -> assertEquals(userId, exc.getUserId()));
+  }
+
   private void mockFindUserById(String userId, String userName) {
-    when(userClient.getUserById(contains(userId))).thenReturn(Optional.of(new UserClient.UserDto(userId, userName)));
+    when(userClient.getUserById(contains(userId))).thenReturn(Optional.of(new User(userId, userName)));
   }
 
   private void mockValidator(PasswordCheck passwordCheck, ValidationErrors errors) {